wsd: generic readFile
